feat: remove the ability to opt out of the new architecture
