+++ boilerplate/app.json
@@ -9,7 +9,7 @@
   "updates": {
     "fallbackToCacheTimeout": 0
   },
-  "newArchEnabled": false,
+  "newArchEnabled": true,
   "jsEngine": "hermes",
   "assetBundlePatterns": [
     "**/*"
